twitter_agent/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
apify_client==2.5.0
apify_shared==2.2.0
beautifulsoup4==4.12.3
certifi==2026.2.25
charset-normalizer==3.4.4
colorama==0.4.6
distro==1.9.0
feedparser==6.0.11
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
impit==0.11.0
jiter==0.13.0
more-itertools==10.8.0
oauthlib==3.3.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
regex==2026.2.28
requests==2.31.0
requests-oauthlib==2.0.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
tavily-python==0.3.3
tiktoken==0.12.0
tqdm==4.67.3
tweepy==4.16.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
